Type application in macro definitions
When a type constructor gets applied to type arguments in a macro definition,
the WeakTypeTag that gets passed into the macro for the resulting type doesn't
seem to have any information about the type arguments.

Why?

Information about the types used as arguments is available: if the macro
definition additionally passes the arguments into the macro, the macro can then
use the types from the WeakTypeTags of the applied type and the arguments to
correctly re-apply the type. But this is really inconvenient, and obscures what
the macro is actually trying to achieve.

This commit demonstrates this phenomenon and a workaround in trivial macros and
accompanying tests.